__init__ will just be there for the language impl
