+++ raco/backends/clang/__init__.py
@@ -1,0 +1,2 @@
+# everything in clang.py made public
+from clang import *
