+++ raco/backends/myria/__init__.py
@@ -1,0 +1,2 @@
+# everything in myrialang made public
+from myrialang import *
