+++ raco/backends/radish/__init__.py
@@ -1,0 +1,2 @@
+# everything in radish made public
+from radish import *
